feat(venue): move NoInformationPlaceholder and NoOfferPlaceholder in new folder /placeholders
