+++ src/features/venue/components/Placeholders/NoInformationPlaceholder.tsx
