+++ src/features/venue/components/Placeholders/NoOfferPlaceholder.tsx
