+++ src/features/venue/components/PracticalInformation.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 
 import { VenueResponse } from 'api/gen'
 import { ContactBlock } from 'features/venue/components/ContactBlock/ContactBlock'
-import { NoInformationPlaceholder } from 'features/venue/components/NoInformationPlaceholder'
+import { NoInformationPlaceholder } from 'features/venue/components/Placeholders/NoInformationPlaceholder'
 import { AccessibilityBlock } from 'ui/components/accessibility/AccessibilityBlock'
 import { Separator } from 'ui/components/Separator'
 import { Spacer, Typo } from 'ui/theme'
