+++ src/features/venue/components/VenueOffers/VenueOffers.tsx
@@ -8,7 +8,7 @@ import { GtlPlaylist } from 'features/gtlPlaylist/components/GtlPlaylist'
 import { useGTLPlaylists } from 'features/gtlPlaylist/hooks/useGTLPlaylists'
 import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useVenueOffers } from 'features/venue/api/useVenueOffers'
-import { NoOfferPlaceholder } from 'features/venue/components/VenueOffers/NoOfferPlaceholder'
+import { NoOfferPlaceholder } from 'features/venue/components/Placeholders/NoOfferPlaceholder'
 import { VenueOfferTile } from 'features/venue/components/VenueOfferTile/VenueOfferTile'
 import { useNavigateToSearchWithVenueOffers } from 'features/venue/helpers/useNavigateToSearchWithVenueOffers'
 import { analytics } from 'libs/analytics'
